feat(mock-page-ui): add collection detail page
