+++ src/app/(client)/collections/[id]/page.tsx
@@ -1,3 +1,6 @@
+import { CollectionsDetail } from "@/components/pages/collections-detail";
+import { popularCollections } from "../../data";
+
 export default async function CollectionDetailPage() {
-  return <div>Collection detail page</div>;
+  return <CollectionsDetail collection={popularCollections[0]} />;
 }
