--- src/app/(client)/collections/layout.tsx
@@ -1,23 +1,0 @@
-import { Header } from "@/components/atoms/header";
-
-type TCollectionLayout = {
-  children: React.ReactNode;
-  searchBox: React.ReactNode;
-};
-
-export default function CollectionLayout({
-  children,
-  searchBox,
-}: TCollectionLayout) {
-  return (
-    <div className="grid grid-rows-[auto_auto_1fr]">
-      <Header className="mt-4" level={1}>
-        Collections List
-      </Header>
-
-      {searchBox}
-
-      {children}
-    </div>
-  );
-}
