+++ src/app/(client)/collections/page.tsx
@@ -1,10 +1,6 @@
-import { CollectionList } from "@/components/organisms/collection-list";
 import { popularCollections } from "../data";
+import { Collections } from "@/components/pages/collections-list";
 
 export default async function CollectionsPage() {
-  return (
-    <div className="py-4">
-      <CollectionList collections={popularCollections} />
-    </div>
-  );
+  return <Collections collections={popularCollections} />;
 }
