+++ src/components/molecules/collection-card/index.tsx
@@ -3,6 +3,8 @@ import React from "react";
 import { LinkAsButton } from "../link-as-button";
 import { useBgImage } from "@/libs/hooks/useBgImage";
 import { joinClass } from "@/helpers/style";
+import { genPath } from "@/helpers/router";
+import { EPath } from "@/constants";
 
 const DEFAULT_CLASSNAME =
   "flex flex-col items-center justify-center gap-4 p-4 border border-gray-200 rounded-lg text-primary-100";
@@ -28,7 +30,7 @@ export const CollectionCard: React.FC<Readonly<TCollectionCard>> = ({
       <Header level={3}>{collection.title}</Header>
       <p>{collection.description}</p>
       <LinkAsButton
-        href={`/collection/${collection.id}`}
+        href={genPath(EPath.collections, collection.id)}
         buttonProps={{
           variant: "outline",
         }}
