--- src/types/collection-card.d.ts
@@ -1,10 +1,0 @@
-type TCollectionCard = {
-  id: string;
-  title: string;
-  description: string;
-  banner: string;
-  products: TProductCard[];
-  className?: string;
-  width?: string;
-  height?: string;
-};
