--- src/types/collection-list.d.ts
@@ -1,4 +1,0 @@
-type TCollectionList = {
-  collections: TCollectionCard[];
-  className?: string;
-};
